+++ app/src/main/java/top/yogiczy/mytv/ui/screens/leanback/settings/SettingsCategories.kt
@@ -4,7 +4,6 @@ import androidx.compose.material.icons.Icons
 import androidx.compose.material.icons.filled.BugReport
 import androidx.compose.material.icons.filled.DisplaySettings
 import androidx.compose.material.icons.filled.FormatListNumbered
-import androidx.compose.material.icons.filled.Http
 import androidx.compose.material.icons.filled.Info
 import androidx.compose.material.icons.filled.LiveTv
 import androidx.compose.material.icons.filled.Menu
@@ -27,7 +26,6 @@ enum class LeanbackSettingsCategories(
     FAVORITE(Icons.Default.Star, "收藏"),
     UPDATE(Icons.Default.Update, "更新"),
     VIDEO_PLAYER(Icons.Default.SmartDisplay, "播放器"),
-    HTTP(Icons.Default.Http, "网络"),
     DEBUG(Icons.Default.BugReport, "调试"),
     LOG(Icons.Default.FormatListNumbered, "日志"),
     MORE(Icons.Default.MoreHoriz, "更多设置"),