+++ app/src/main/java/top/yogiczy/mytv/ui/screens/leanback/settings/components/SettingsCategoryContent.kt
@@ -32,7 +32,6 @@ fun LeanbackSettingsCategoryContent(
             LeanbackSettingsCategories.FAVORITE -> LeanbackSettingsCategoryFavorite()
             LeanbackSettingsCategories.UPDATE -> LeanbackSettingsCategoryUpdate()
             LeanbackSettingsCategories.VIDEO_PLAYER -> LeanbackSettingsCategoryVideoPlayer()
-            LeanbackSettingsCategories.HTTP -> LeanbackSettingsCategoryHttp()
             LeanbackSettingsCategories.DEBUG -> LeanbackSettingsCategoryDebug()
             LeanbackSettingsCategories.LOG -> LeanbackSettingsCategoryLog(
                 history = Logger.history,