+++ app/src/main/java/top/yogiczy/mytv/ui/screens/leanback/settings/components/SettingsCategoryVideoPlayer.kt
@@ -73,7 +73,7 @@ fun LeanbackSettingsCategoryVideoPlayer(
 
 @Preview
 @Composable
-private fun LeanbackSettingsCategoryHttpPreview() {
+private fun LeanbackSettingsCategoryVideoPlayerPreview() {
     SP.init(LocalContext.current)
     LeanbackTheme {
         LeanbackSettingsCategoryVideoPlayer(